Dumps statistics every day at midnight and noon
